shreyave-code/requirements.txt
Pillow>=11.0.0
python-dotenv>=1.0.0
